@@ -1,17 +1,18 @@
 {
-  "name": "polis-react",
+  "name": "polis-client-admin",
   "version": "0.0.1",
   "repository": "github:pol-is/polisClientAdmin",
-  "description": "Polis React",
+  "description": "Polis Admin Console",
   "engines": {
-    "node": "6.10.3"
+    "node": "8.17.0"
   },
   "scripts": {
     "clean": "rimraf dist",
     "build:webpack": "npm run clean && NODE_ENV=production webpack --config webpack.config.js",
-    "build:webpack_unminified": "npm run clean && NODE_ENV=production webpack --config webpack.config.unminified.js",
+    "build:webpack_unminified": "npm run clean && NODE_ENV=development webpack --config webpack.config.unminified.js",
     "build": "npm run build:webpack",
     "build:preprod": "gulp deployPreprod",
+    "build-debug": "npm run build:webpack_unminified",
     "deploy:prod": "./deploy_TO_PRODUCTION",
     "deploy:preprod": "./deployPreprod",
     "start": "./x",
@@ -54,46 +59,46 @@
     "rimraf": "^2.4.3",
     "run-sequence": "~0.3.6",
     "scp2": "^0.5.0",
-    "webpack": "^1.12.2",
-    "webpack-dev-middleware": "^1.2.0",
-    "webpack-hot-middleware": "^2.4.1"
+    "webpack": "^4.43.0",
+    "webpack-bundle-analyzer": "^3.8.0",
+    "webpack-cli": "^3.3.11",
+    "webpack-dev-middleware": "^1.2.0"
   },
   "dependencies": {
+    "@babel/runtime": "^7.9.6",
+    "@material-ui/core": "^4.9.14",
     "boolean": "^0.1.3",
     "color": "^0.7.3",
+    "d3": "^5.16.0",
     "highlight.js": "^8.9.1",
-    "history": "^1.12.5",
+    "history": "4.10.1",
     "jquery": "^2.1.4",
     "lodash": "^3.10.1",
-    "material-ui": "^0.13.4",
+    "mime": "^1.6.0",
     "moment": "^2.10.6",
-    "prop-types": "^15.5.10",
-    "radium": "^0.15.0",
-    "react": "^15.5.4",
-    "react-codemirror": "^0.2.1",
-    "react-dom": "^15.5.4",
-    "react-fontawesome": "^0.3.3",
-    "react-highlight": "^0.5.1",
-    "react-markdown": "^1.1.0",
-    "react-redux": "^4.0.0",
-    "react-router": "^2.0.0-rc4",
-    "react-s-alert": "^1.3.0",
-    "react-sidebar": "^2.0.0",
-    "react-stripe-elements": "0.0.2",
-    "react-tap-event-plugin": "^0.2.1",
+    "natives": "^1.1.6",
+    "prop-types": "^15.7.2",
+    "radium": "0.20.1",
+    "react": "^16.4.0",
+    "react-addons-pure-render-mixin": "^15.6.2",
+    "react-addons-update": "^15.6.2",
+    "react-dom": "^16.4.0",
+    "react-easy-emoji": "^1.4.0",
+    "react-fontawesome": "^1.7.1",
+    "react-highlight": "0.12.0",
+    "react-markdown": "4.3.1",
+    "react-redux": "7.2.0",
+    "react-router": "5.2.0",
+    "react-router-dom": "^5.2.0",
+    "react-tap-event-plugin": "^2.0.0",
+    "react-transition-group": "^4.4.1",
     "redux": "^3.0.2",
     "redux-actions": "^0.9.0",
-    "redux-devtools": "^2.1.5",
     "redux-logger": "^2.0.4",
-    "redux-router": "^1.0.0-beta3",
     "redux-thunk": "^1.0.0",
     "request": "^2.65.0",
-    "tachyons-react-social-icons": "^2.8.4",
-    "victory-axis": "1.5.7",
-    "victory-bar": "2.5.0",
-    "victory-chart": "2.1.1",
-    "victory-line": "0.5.6",
-    "victory-pie": "0.1.4",
+    "theme-ui": "^0.3.1",
+    "victory": "^34.3.11",
     "whatwg-fetch": "^0.10.1"
   },
   "jest": {
