@@ -1,7 +1,7 @@
 // Copyright (C) 2012-present, The Authors. This program is free software: you can redistribute it and/or  modify it under the terms of the GNU Affero General Public License, version 3, as published by the Free Software Foundation. This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for more details. You should have received a copy of the GNU Affero General Public License along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import Radium from 'radium';
-import React from 'react';
+import Radium from "radium";
+import React from "react";
 
 var styles = {
   base: {
@@ -15,26 +15,26 @@ var styles = {
     cursor: "pointer",
     outline: "none",
 
-    '@media (min-width: 992px)': {
-      padding: "0.6em 1.2em"
+    "@media (min-width: 992px)": {
+      padding: "0.6em 1.2em",
     },
 
-    '@media (min-width: 1200px)': {
-      padding: "0.8em 1.5em"
+    "@media (min-width: 1200px)": {
+      padding: "0.8em 1.5em",
     },
 
-    ':hover': {
-      backgroundColor: "#0088FF"
+    ":hover": {
+      backgroundColor: "#0088FF",
     },
 
-    ':focus': {
-      backgroundColor: "#0088FF"
+    ":focus": {
+      backgroundColor: "#0088FF",
     },
 
-    ':active': {
+    ":active": {
       backgroundColor: "#005299",
       transform: "translateY(2px)",
-    }
+    },
   },
 };
 
@@ -42,21 +42,11 @@ var styles = {
 class Button extends React.Component {
   render() {
     return (
-      <button
-        onClick={this.props.onClick}
-        style={[
-          styles.base,
-          this.props.style
-        ]}>
+      <button onClick={this.props.onClick} style={[styles.base, this.props.style]}>
         {this.props.children}
       </button>
     );
   }
 }
 
-Button.propTypes = {
-  onClick: React.PropTypes.func
-};
-
-
-module.exports = Button;
+export default Button;
