@@ -3,7 +3,6 @@
 import $ from "jquery";
 import PolisNet from "../util/net";
 
-
 /* ======= Types ======= */
 export const REQUEST_USER = "REQUEST_USER";
 export const RECEIVE_USER = "RECEIVE_USER";
@@ -164,14 +163,14 @@ export const MATH_FETCH_ERROR = "MATH_FETCH_ERROR";
 
 const requestUser = () => {
   return {
-    type: REQUEST_USER
+    type: REQUEST_USER,
   };
 };
 
 const receiveUser = (data) => {
   return {
     type: RECEIVE_USER,
-    data: data
+    data: data,
   };
 };
 
@@ -179,30 +178,29 @@ const userFetchError = (err) => {
   return {
     type: USER_FETCH_ERROR,
     status: err.status,
-    data: err
-  }
-}
+    data: err,
+  };
+};
 
 const fetchUser = () => {
-  return PolisNet.polisGet("/api/v3/users", {errIfNoAuth: true});
-}
+  return PolisNet.polisGet("/api/v3/users", { errIfNoAuth: true });
+};
 
 export const populateUserStore = () => {
   return (dispatch) => {
-    dispatch(requestUser())
+    dispatch(requestUser());
     return fetchUser().then(
-      res => dispatch(receiveUser(res)),
-      err => dispatch(userFetchError(err))
-    )
-  }
-}
-
+      (res) => dispatch(receiveUser(res)),
+      (err) => dispatch(userFetchError(err))
+    );
+  };
+};
 
 /* signin */
 
 const signinInitiated = () => {
   return {
-    type: SIGNIN_INITIATED
+    type: SIGNIN_INITIATED,
   };
 };
 
@@ -211,7 +209,7 @@ const signinInitiated = () => {
 const signinError = (err) => {
   return {
     type: SIGNIN_ERROR,
-    data: err
+    data: err,
   };
 };
 
@@ -219,7 +217,7 @@ const signinPost = (attrs) => {
   return PolisNet.polisPost("/api/v3/auth/login", attrs);
 };
 
-export const doSignin = (attrs, dest) => {
+export const doSignin = (attrs) => {
   return (dispatch) => {
     dispatch(signinInitiated());
     return signinPost(attrs).then(
@@ -227,7 +225,8 @@ export const doSignin = (attrs, dest) => {
         setTimeout(() => {
           // Force page to load so we can be sure the password is cleared from memory
           // delay a bit so the cookie has time to set
-          window.location = dest||"";
+          dispatch({ type: "signin completed successfully" });
+          window.location = "/";
         }, 3000);
       },
       (err) => dispatch(signinError(err))
@@ -239,7 +238,7 @@ export const doSignin = (attrs, dest) => {
 
 const createUserInitiated = () => {
   return {
-    type: CREATEUSER_INITIATED
+    type: CREATEUSER_INITIATED,
   };
 };
 
@@ -248,7 +247,7 @@ const createUserInitiated = () => {
 const createUserError = (err) => {
   return {
     type: CREATEUSER_ERROR,
-    data: err
+    data: err,
   };
 };
 
@@ -264,7 +263,7 @@ export const doCreateUser = (attrs, dest) => {
         setTimeout(() => {
           // Force page to load so we can be sure the password is cleared from memory
           // delay a bit so the cookie has time to set
-          window.location = dest||"";
+          window.location = dest || "";
         }, 3000);
       },
       (err) => dispatch(createUserError(err))
@@ -276,20 +275,20 @@ export const doCreateUser = (attrs, dest) => {
 
 const passwordResetInitInitiated = () => {
   return {
-    type: PWRESET_INIT_INITIATED
+    type: PWRESET_INIT_INITIATED,
   };
 };
 
 const passwordResetInitSuccess = () => {
   return {
-    type: PWRESET_INIT_SUCCESS
+    type: PWRESET_INIT_SUCCESS,
   };
 };
 
 const passwordResetInitError = (err) => {
   return {
     type: PWRESET_INIT_ERROR,
-    data: err
+    data: err,
   };
 };
 
@@ -319,20 +318,20 @@ export const doPasswordResetInit = (attrs) => {
 
 const passwordResetInitiated = () => {
   return {
-    type: PWRESET_INITIATED
+    type: PWRESET_INITIATED,
   };
 };
 
 const passwordResetSuccess = () => {
   return {
-    type: PWRESET_SUCCESS
+    type: PWRESET_SUCCESS,
   };
 };
 
 const passwordResetError = (err) => {
   return {
     type: PWRESET_ERROR,
-    data: err
+    data: err,
   };
 };
 
@@ -362,20 +361,20 @@ export const doPasswordReset = (attrs) => {
 
 const facebookSigninInitiated = () => {
   return {
-    type: FACEBOOK_SIGNIN_INITIATED
+    type: FACEBOOK_SIGNIN_INITIATED,
   };
 };
 
 const facebookSigninSuccessful = () => {
   return {
-    type: FACEBOOK_SIGNIN_SUCCESSFUL
+    type: FACEBOOK_SIGNIN_SUCCESSFUL,
   };
 };
 
 const facebookSigninFailed = (errorCode) => {
   return {
     type: FACEBOOK_SIGNIN_FAILED,
-    errorCode: errorCode
+    errorCode: errorCode,
   };
 };
 
@@ -383,7 +382,6 @@ const getFriends = () => {
   var dfd = $.Deferred();
 
   const getMoreFriends = (friendsSoFar, urlForNextCall) => {
-
     return $.get(urlForNextCall).then((response) => {
       if (response.data.length) {
         for (let i = 0; i < response.data.length; i++) {
@@ -397,16 +395,13 @@ const getFriends = () => {
         return friendsSoFar;
       }
     });
-  }
+  };
 
   FB.api("/me/friends", (response) => {
     if (response && !response.error) {
-
       var friendsSoFar = response.data;
       if (response.data.length && response.paging.next) {
-        getMoreFriends(friendsSoFar, response.paging.next).then(
-          dfd.resolve,
-          dfd.reject);
+        getMoreFriends(friendsSoFar, response.paging.next).then(dfd.resolve, dfd.reject);
       } else {
         dfd.resolve(friendsSoFar || []);
       }
@@ -416,7 +411,7 @@ const getFriends = () => {
     }
   });
   return dfd.promise();
-}
+};
 
 const getInfo = () => {
   var dfd = $.Deferred();
@@ -441,8 +436,6 @@ const getInfo = () => {
     if (response && !response.error) {
       if (response.location && response.location.id) {
         FB.api("/" + response.location.id, (locationResponse) => {
-          // console.log("locationResponse");
-          // console.dir(locationResponse);
           if (locationResponse) {
             response.locationInfo = locationResponse;
           }
@@ -457,61 +450,60 @@ const getInfo = () => {
     }
   });
   return dfd.promise();
-}
+};
 
 const saveFacebookFriendsData = (data, dest, dispatch) => {
   $.ajax({
     url: "/api/v3/auth/facebook",
     contentType: "application/json; charset=utf-8",
     headers: {
-      "Cache-Control": "max-age=0"
+      "Cache-Control": "max-age=0",
     },
     xhrFields: {
-        withCredentials: true
+      withCredentials: true,
     },
     dataType: "json",
     data: JSON.stringify(data),
-    type: "POST"
-  }).then(() => {
-    setTimeout(() => {
+    type: "POST",
+  }).then(
+    () => {
+      setTimeout(() => {
         // Force page to load so we can be sure the old user"s state is cleared from memory
         // delay a bit so the cookies have time to clear too.
         window.location = dest || "/";
       }, 1000);
-    }, (err) => {
-    console.dir(err);
+    },
+    (err) => {
+      console.dir(err);
 
-    if ( err.responseText && /polis_err_user_with_this_email_exists/.test(err.responseText) ) {
+      if (err.responseText && /polis_err_user_with_this_email_exists/.test(err.responseText)) {
+        // Todo handle
 
-      console.log("thats a user already enter your password")
-      // Todo handle
+        // var password = prompt("A pol.is user "+data.fb_email+", the same email address as associted with your facebook account, already exists. Enter your pol.is password to enable facebook login for your pol.is account.");
+        // that.linkMode = true;
 
-      // var password = prompt("A pol.is user "+data.fb_email+", the same email address as associted with your facebook account, already exists. Enter your pol.is password to enable facebook login for your pol.is account.");
-      // that.linkMode = true;
+        dispatch(facebookSigninFailed("polis_err_user_with_this_email_exists")); //handle case user already exists enter your password
 
-      dispatch(facebookSigninFailed("polis_err_user_with_this_email_exists")) //handle case user already exists enter your password
-
-      // that.model.set({
-      //   create: false, // don"t show create account stuff, account exists.
-      //   linkMode: true,
-      //   email: data.fb_email,
-      // });
-    } else {
-      alert("error logging in with Facebook");
+        // that.model.set({
+        //   create: false, // don"t show create account stuff, account exists.
+        //   linkMode: true,
+        //   email: data.fb_email,
+        // });
+      } else {
+        alert("error logging in with Facebook");
+      }
     }
-  });
+  );
 };
 
 const processFacebookFriendsData = (response, dest, dispatch, optionalPassword) => {
-
   return (fb_public_profile, friendsData) => {
     // alert(JSON.stringify(friendsData));
-    console.log("got info and friends");
 
     let data = {
       fb_public_profile: JSON.stringify(fb_public_profile),
       fb_friends_response: JSON.stringify(friendsData),
-      response: JSON.stringify(response)
+      response: JSON.stringify(response),
     };
 
     // cleaner as fb_email: fb_public_profile.email ? fb_public_profile.email : null
@@ -522,10 +514,7 @@ const processFacebookFriendsData = (response, dest, dispatch, optionalPassword) 
       data.provided_email = prompt("Please enter your email address.");
     }
 
-    let hname = [
-      fb_public_profile.first_name,
-      fb_public_profile.last_name
-    ].join(" ");
+    let hname = [fb_public_profile.first_name, fb_public_profile.last_name].join(" ");
 
     if (hname.length) {
       data.hname = hname;
@@ -539,59 +528,53 @@ const processFacebookFriendsData = (response, dest, dispatch, optionalPassword) 
       data.password = optionalPassword;
     }
 
-    saveFacebookFriendsData(data, dest, dispatch)
-  }
-
-}
+    saveFacebookFriendsData(data, dest, dispatch);
+  };
+};
 
 const onFbLoginOk = (response, dest, dispatch, optionalPassword) => {
-
-  console.log("onFbLoginOk");
-  console.dir(response);
-
-  $.when(
-    getInfo(),
-    getFriends()
-  ).then(
+  $.when(getInfo(), getFriends()).then(
     processFacebookFriendsData(response, dest, dispatch, optionalPassword),
     (err) => {
       console.error(err);
       console.dir(arguments);
     }
   );
-}
+};
 
 const callFacebookLoginAPI = (dest, dispatch, optionalPassword) => {
   console.log("ringing facebook...");
-  const password = "THIS_STRING_SHOULD_NOT_BE_HERE"
+  const password = "THIS_STRING_SHOULD_NOT_BE_HERE";
 
-  FB.login((res) =>  {
+  FB.login(
+    (res) => {
       return onFbLoginOk(res, dest, dispatch, optionalPassword);
-    }, {
+    },
+    {
       return_scopes: true, // response should contain the scopes the user allowed
       scope: [
         // "taggable_friends", // requires review.
         // invitable_friends NOTE: only for games with a fb Canvas presence, so don"t use this
         "public_profile",
         "user_friends",
-        "email"
-      ].join(",")
-    });
-}
+        "email",
+      ].join(","),
+    }
+  );
+};
 
 export const doFacebookSignin = (dest, optionalPassword) => {
   return (dispatch) => {
-    dispatch(facebookSigninInitiated())
-    console.log("facebook sign in initiated", dest)
-    return callFacebookLoginAPI(dest, dispatch, optionalPassword)
-  }
-}
+    dispatch(facebookSigninInitiated());
+    return callFacebookLoginAPI(dest, dispatch, optionalPassword);
+  };
+};
 
 /* signout */
 
 const signoutInitiated = () => {
   return {
-    type: SIGNOUT_INITIATED
+    type: SIGNOUT_INITIATED,
   };
 };
 
@@ -600,9 +583,9 @@ const signoutInitiated = () => {
 const signoutError = (err) => {
   return {
     type: SIGNOUT_ERROR,
-    data: err
-  }
-}
+    data: err,
+  };
+};
 
 const signoutPost = (dest) => {
   // relying on server to clear cookies
@@ -611,60 +594,60 @@ const signoutPost = (dest) => {
     url: "/api/v3/auth/deregister",
     data: {},
     dataType: "text", // server returns an empty response, so can"t parse as JSON
-    });
-}
+  });
+};
 
 export const doSignout = (dest) => {
   return (dispatch) => {
     dispatch(signoutInitiated());
     return signoutPost().then(
-      res => {
+      (res) => {
         setTimeout(() => {
           // Force page to load so we can be sure the old user"s state is cleared from memory
           // delay a bit so the cookies have time to clear too.
-          window.location = dest || "/about";
+          window.location = dest || "/home";
         }, 1000);
       },
-      err => dispatch(signoutError(err))
-    )
-  }
-}
+      (err) => dispatch(signoutError(err))
+    );
+  };
+};
 
 /* Conversations */
 
 const requestConversations = () => {
   return {
-    type: REQUEST_CONVERSATIONS
+    type: REQUEST_CONVERSATIONS,
   };
 };
 
 const receiveConversations = (data) => {
   return {
     type: RECEIVE_CONVERSATIONS,
-    data: data
+    data: data,
   };
 };
 
 const conversationsError = (err) => {
   return {
     type: CONVERSATIONS_FETCH_ERROR,
-    data: err
-  }
-}
+    data: err,
+  };
+};
 
 const fetchConversations = () => {
   return $.get("/api/v3/conversations?include_all_conversations_i_am_in=true");
-}
+};
 
 export const populateConversationsStore = () => {
   return (dispatch) => {
-    dispatch(requestConversations())
+    dispatch(requestConversations());
     return fetchConversations().then(
-      res => dispatch(receiveConversations(res)),
-      err => dispatch(conversationsError(err))
-    )
-  }
-}
+      (res) => dispatch(receiveConversations(res)),
+      (err) => dispatch(conversationsError(err))
+    );
+  };
+};
 
 /* zid metadata */
 
@@ -672,47 +655,47 @@ const requestZidMetadata = (conversation_id) => {
   return {
     type: REQUEST_ZID_METADATA,
     data: {
-      conversation_id: conversation_id
-    }
+      conversation_id: conversation_id,
+    },
   };
 };
 
 const receiveZidMetadata = (data) => {
   return {
     type: RECEIVE_ZID_METADATA,
-    data: data
+    data: data,
   };
 };
 
 const zidMetadataFetchError = (err) => {
   return {
     type: ZID_METADATA_FETCH_ERROR,
-    data: err
-  }
-}
+    data: err,
+  };
+};
 
 export const resetMetadataStore = () => {
   return {
-    type: ZID_METADATA_RESET
-  }
-}
+    type: ZID_METADATA_RESET,
+  };
+};
 
 const fetchZidMetadata = (conversation_id) => {
   return $.get("/api/v3/conversations?conversation_id=" + conversation_id);
-}
+};
 
 export const populateZidMetadataStore = (conversation_id) => {
   return (dispatch, getState) => {
-
     var state = getState();
-    var hasConversationId = state.zid_metadata &&
+    var hasConversationId =
+      state.zid_metadata &&
       state.zid_metadata.zid_metadata &&
       state.zid_metadata.zid_metadata.conversation_id;
 
     var isLoading = state.zid_metadata.loading;
     // NOTE: if there are multiple calls outstanding this may be wrong.
-    var isLoadingThisConversation = (state.zid_metadata.conversation_id == conversation_id) && isLoading;
-
+    var isLoadingThisConversation =
+      state.zid_metadata.conversation_id == conversation_id && isLoading;
 
     if (isLoadingThisConversation) {
       return;
@@ -723,35 +706,35 @@ export const populateZidMetadataStore = (conversation_id) => {
       return;
     }
 
-    dispatch(requestZidMetadata(conversation_id))
+    dispatch(requestZidMetadata(conversation_id));
     return fetchZidMetadata(conversation_id).then(
-      res => dispatch(receiveZidMetadata(res)),
-      err => dispatch(zidMetadataFetchError(err))
+      (res) => dispatch(receiveZidMetadata(res)),
+      (err) => dispatch(zidMetadataFetchError(err))
     );
-  }
-}
+  };
+};
 
 /* zid metadata update */
 
 const updateZidMetadataStarted = () => {
   return {
-    type: UPDATE_ZID_METADATA_STARTED
-  }
-}
+    type: UPDATE_ZID_METADATA_STARTED,
+  };
+};
 
 const updateZidMetadataSuccess = (data) => {
   return {
     type: UPDATE_ZID_METADATA_SUCCESS,
-    data: data
-  }
-}
+    data: data,
+  };
+};
 
 const updateZidMetadataError = (err) => {
   return {
     type: UPDATE_ZID_METADATA_ERROR,
-    data: err
-  }
-}
+    data: err,
+  };
+};
 
 const updateZidMetadata = (zm, field, value) => {
   let data = {};
@@ -763,54 +746,55 @@ const updateZidMetadata = (zm, field, value) => {
     headers: { "Cache-Control": "max-age=0" },
     xhrFields: { withCredentials: true },
     dataType: "json",
-    data: JSON.stringify(Object.assign({}, zm, data))
-  })
-}
+    data: JSON.stringify(Object.assign({}, zm, data)),
+  });
+};
 
 export const handleZidMetadataUpdate = (zm, field, value) => {
   return (dispatch) => {
-    dispatch(updateZidMetadataStarted())
+    dispatch(updateZidMetadataStarted());
     return updateZidMetadata(zm, field, value)
-      .then(res => dispatch(updateZidMetadataSuccess(res)))
-      .fail(err => dispatch(updateZidMetadataError(err)))
-  }
-}
+      .then((res) => dispatch(updateZidMetadataSuccess(res)))
+      .fail((err) => dispatch(updateZidMetadataError(err)));
+  };
+};
 
 export const optimisticZidMetadataUpdateOnTyping = (zm, field, value) => {
-  zm[field] = value
+  zm[field] = value;
   return {
     type: OPTIMISTIC_ZID_METADATA_UPDATE,
-    data: zm
-  }
-}
+    data: zm,
+  };
+};
 
 /* seed comments submit */
 
 export const seedCommentChanged = (text) => {
   return {
     type: SEED_COMMENT_LOCAL_UPDATE,
-    text: text
-  }
-}
+    text: text,
+  };
+};
 
 const submitSeedCommentStart = () => {
   return {
-    type: SUBMIT_SEED_COMMENT
-  }
-}
+    type: SUBMIT_SEED_COMMENT,
+  };
+};
 
 const submitSeedCommentPostSuccess = () => {
+  console.log("seed comment post success");
   return {
-    type: SUBMIT_SEED_COMMENT_SUCCESS
-  }
-}
+    type: SUBMIT_SEED_COMMENT_SUCCESS,
+  };
+};
 
 const submitSeedCommentPostError = (err) => {
   return {
     type: SUBMIT_SEED_COMMENT_ERROR,
     data: err,
-  }
-}
+  };
+};
 
 const postSeedComment = (comment) => {
   return PolisNet.polisPost("/api/v3/comments", comment);
@@ -819,12 +803,12 @@ const postSeedComment = (comment) => {
 export const handleSeedCommentSubmit = (comment) => {
   return (dispatch) => {
     dispatch(submitSeedCommentStart());
-    return postSeedComment(comment).then(
-      (res) => dispatch(submitSeedCommentPostSuccess(res)),
-      (err) => dispatch(submitSeedCommentPostError(err))
-    ).then(dispatch(
-      populateAllCommentStores(comment.conversation_id)
-    ));
+    return postSeedComment(comment)
+      .then(
+        (res) => dispatch(submitSeedCommentPostSuccess(res)),
+        (err) => dispatch(submitSeedCommentPostError(err))
+      )
+      .then(dispatch(populateAllCommentStores(comment.conversation_id)));
   };
 };
 
@@ -837,57 +821,57 @@ const makeStandardError = (type, err) => {
   return {
     type: type,
     data: err,
-  }
+  };
 };
 const makeStandardSuccess = (type, data) => {
   return {
     type: type,
     data: data,
-  }
+  };
 };
-
-
 
 /* seed tweets submit */
 
 export const seedCommentTweetChanged = (text) => {
   return {
     type: SEED_COMMENT_TWEET_LOCAL_UPDATE,
-    text: text
-  }
-}
+    text: text,
+  };
+};
 const submitSeedCommentTweetStart = () => {
   return {
-    type: SUBMIT_SEED_COMMENT_TWEET
-  }
-}
+    type: SUBMIT_SEED_COMMENT_TWEET,
+  };
+};
 
 const submitSeedCommentPostTweetSuccess = () => {
   return {
-    type: SUBMIT_SEED_COMMENT_TWEET_SUCCESS
-  }
-}
+    type: SUBMIT_SEED_COMMENT_TWEET_SUCCESS,
+  };
+};
 
 const submitSeedCommentPostTweetError = (err) => {
   return {
     type: SUBMIT_SEED_COMMENT_TWEET_ERROR,
     data: err,
-  }
-}
+  };
+};
 
 const postSeedCommentTweet = (o) => {
-  return PolisNet.polisPost('/api/v3/comments', o);
-}
+  return PolisNet.polisPost("/api/v3/comments", o);
+};
 
 export const handleSeedCommentTweetSubmit = (o) => {
   return (dispatch) => {
-    dispatch(submitSeedCommentTweetStart())
-    return postSeedCommentTweet(o).then(
-      res => dispatch(submitSeedCommentPostTweetSuccess(res)),
-      err => dispatch(submitSeedCommentPostTweetError(err))
-    ).then(dispatch(populateAllCommentStores(o.conversation_id)))
-  }
-}
+    dispatch(submitSeedCommentTweetStart());
+    return postSeedCommentTweet(o)
+      .then(
+        (res) => dispatch(submitSeedCommentPostTweetSuccess(res)),
+        (err) => dispatch(submitSeedCommentPostTweetError(err))
+      )
+      .then(dispatch(populateAllCommentStores(o.conversation_id)));
+  };
+};
 
 /* seed comments fetch */
 
@@ -925,79 +909,83 @@ export const handleSeedCommentTweetSubmit = (o) => {
 //   }
 // }
 
-
-
 /* create conversation */
 
 const createConversationStart = () => {
   return {
-    type: CREATE_NEW_CONVERSATION
-  }
-}
+    type: CREATE_NEW_CONVERSATION,
+  };
+};
 
 const createConversationPostSuccess = (res) => {
   return {
     type: CREATE_NEW_CONVERSATION_SUCCESS,
     data: res,
-  }
-}
+  };
+};
 
 const createConversationPostError = (err) => {
   return {
     type: CREATE_NEW_CONVERSATION_ERROR,
     data: err,
-  }
-}
+  };
+};
 
 const postCreateConversation = () => {
   return PolisNet.polisPost("/api/v3/conversations", {
     is_draft: true,
     is_active: true,
   });
-}
-
-export const handleCreateConversationSubmit = (routeTo) => {
-  return (dispatch, getState) => {
-    dispatch(createConversationStart())
-    return postCreateConversation().then(
-      (res) => {
-        dispatch(createConversationPostSuccess(res));
-        return res;
-      },
-      (err) => dispatch(createConversationPostError(err))
-    ).then((res) => {
-      window.location = "/m/" + res.conversation_id;
-    });
-  };
 };
 
+export const handleCreateConversationSubmit = (routeTo) => {
+  return (dispatch) => {
+    dispatch(createConversationStart());
+    return postCreateConversation()
+      .then(
+        (res) => {
+          dispatch(createConversationPostSuccess(res));
+          return res;
+        },
+        (err) => dispatch(createConversationPostError(err))
+      )
+      .then((res) => {
+        window.location = "/m/" + res.conversation_id;
+      });
+  };
+};
 
 /* request all comments */
 
 const requestComments = () => {
   return {
-    type: REQUEST_COMMENTS
-  }
+    type: REQUEST_COMMENTS,
+  };
 };
 
 const receiveComments = (data) => {
   return {
     type: RECEIVE_COMMENTS,
-    data: data
-  }
+    data: data,
+  };
 };
 
 const commentsFetchError = (err) => {
   return {
     type: COMMENTS_FETCH_ERROR,
-    data: err
-  }
-}
+    data: err,
+  };
+};
 
 const fetchAllComments = (conversation_id) => {
   // let includeSocial = "include_social=true&";
   let includeSocial = "";
-  return $.get("/api/v3/comments?moderation=true&include_voting_patterns=false&"+includeSocial+"conversation_id=" + conversation_id);
+  return $.get(
+    "/api/v3/comments?moderation=true&include_voting_patterns=false&" +
+      includeSocial +
+      "conversation_id=" +
+      conversation_id
+  );
 };
 
 export const populateCommentsStore = (conversation_id) => {
@@ -1014,32 +1002,32 @@ export const populateCommentsStore = (conversation_id) => {
 
 const requestMath = () => {
   return {
-    type: REQUEST_MATH
+    type: REQUEST_MATH,
   };
 };
 
 const receiveMath = (data) => {
   return {
     type: RECEIVE_MATH,
-    data: data
-  }
+    data: data,
+  };
 };
 
 const mathFetchError = (err) => {
   return {
     type: MATH_FETCH_ERROR,
-    data: err
+    data: err,
   };
 };
 
 const fetchMath = (conversation_id, math_tick) => {
-  return $.get("/api/v3/math/pca2?&math_tick="+ math_tick +"&conversation_id=" + conversation_id);
+  return $.get("/api/v3/math/pca2?&math_tick=" + math_tick + "&conversation_id=" + conversation_id);
 };
 
 export const populateMathStore = (conversation_id) => {
   return (dispatch, getState) => {
     dispatch(requestMath());
-    const math_tick = getState().math.math_tick
+    const math_tick = getState().math.math_tick;
     return fetchMath(conversation_id, math_tick).then(
       (res) => dispatch(receiveMath(res)),
       (err) => dispatch(mathFetchError(err))
@@ -1058,32 +1046,37 @@ const requestUnmoderatedComments = () => {
 const receiveUnmoderatedComments = (data) => {
   return {
     type: RECEIVE_UNMODERATED_COMMENTS,
-    data: data
+    data: data,
   };
 };
 
 const unmoderatedCommentsFetchError = (err) => {
   return {
     type: UNMODERATED_COMMENTS_FETCH_ERROR,
-    data: err
-  }
-}
+    data: err,
+  };
+};
 
 const fetchUnmoderatedComments = (conversation_id) => {
   // let includeSocial = "include_social=true&";
   let includeSocial = "";
-  return $.get("/api/v3/comments?moderation=true&include_voting_patterns=false&" +includeSocial+"mod=0&conversation_id=" + conversation_id);
-}
+  return $.get(
+    "/api/v3/comments?moderation=true&include_voting_patterns=false&" +
+      includeSocial +
+      "mod=0&conversation_id=" +
+      conversation_id
+  );
+};
 
 export const populateUnmoderatedCommentsStore = (conversation_id) => {
   return (dispatch) => {
-    dispatch(requestUnmoderatedComments())
+    dispatch(requestUnmoderatedComments());
     return fetchUnmoderatedComments(conversation_id).then(
-      res => dispatch(receiveUnmoderatedComments(res)),
-      err => dispatch(unmoderatedCommentsFetchError(err))
-    )
-  }
-}
+      (res) => dispatch(receiveUnmoderatedComments(res)),
+      (err) => dispatch(unmoderatedCommentsFetchError(err))
+    );
+  };
+};
 
 /* accepted comments */
 
@@ -1096,32 +1089,37 @@ const requestAcceptedComments = () => {
 const receiveAcceptedComments = (data) => {
   return {
     type: RECEIVE_ACCEPTED_COMMENTS,
-    data: data
+    data: data,
   };
 };
 
 const acceptedCommentsFetchError = (err) => {
   return {
     type: ACCEPTED_COMMENTS_FETCH_ERROR,
-    data: err
-  }
-}
+    data: err,
+  };
+};
 
 const fetchAcceptedComments = (conversation_id) => {
   // let includeSocial = "include_social=true&";
   let includeSocial = "";
-  return $.get("/api/v3/comments?moderation=true&include_voting_patterns=false&mod=1&"+includeSocial+"conversation_id=" + conversation_id);
-}
+  return $.get(
+    "/api/v3/comments?moderation=true&include_voting_patterns=false&mod=1&" +
+      includeSocial +
+      "conversation_id=" +
+      conversation_id
+  );
+};
 
 export const populateAcceptedCommentsStore = (conversation_id) => {
   return (dispatch) => {
-    dispatch(requestAcceptedComments())
+    dispatch(requestAcceptedComments());
     return fetchAcceptedComments(conversation_id).then(
-      res => dispatch(receiveAcceptedComments(res)),
-      err => dispatch(acceptedCommentsFetchError(err))
-    )
-  }
-}
+      (res) => dispatch(receiveAcceptedComments(res)),
+      (err) => dispatch(acceptedCommentsFetchError(err))
+    );
+  };
+};
 
 /* rejected comments */
 
@@ -1134,30 +1132,34 @@ const requestRejectedComments = () => {
 const receiveRejectedComments = (data) => {
   return {
     type: RECEIVE_REJECTED_COMMENTS,
-    data: data
+    data: data,
   };
 };
 
 const rejectedCommentsFetchError = (err) => {
   return {
     type: REJECTED_COMMENTS_FETCH_ERROR,
-    data: err
-  }
-}
-
+    data: err,
+  };
+};
 
 const fetchRejectedComments = (conversation_id) => {
   // let includeSocial = "include_social=true&";
   let includeSocial = "";
-  return $.get("/api/v3/comments?moderation=true&include_voting_patterns=false&"+includeSocial+"mod=-1&conversation_id=" + conversation_id);
-}
+  return $.get(
+    "/api/v3/comments?moderation=true&include_voting_patterns=false&" +
+      includeSocial +
+      "mod=-1&conversation_id=" +
+      conversation_id
+  );
+};
 
 export const populateRejectedCommentsStore = (conversation_id) => {
   return (dispatch) => {
-    dispatch(requestRejectedComments())
+    dispatch(requestRejectedComments());
     return fetchRejectedComments(conversation_id).then(
-      res => dispatch(receiveRejectedComments(res)),
-      err => dispatch(rejectedCommentsFetchError(err))
+      (res) => dispatch(receiveRejectedComments(res)),
+      (err) => dispatch(rejectedCommentsFetchError(err))
     );
   };
 };
@@ -1182,88 +1184,87 @@ export const populateAllCommentStores = (conversation_id) => {
 const optimisticCommentAccepted = (comment) => {
   return {
     type: ACCEPT_COMMENT,
-    comment: comment
-  }
-}
+    comment: comment,
+  };
+};
 
 const acceptCommentSuccess = (data) => {
   return {
     type: ACCEPT_COMMENT_SUCCESS,
-    data: data
-  }
-}
+    data: data,
+  };
+};
 
 const acceptCommentError = (err) => {
   return {
     type: ACCEPT_COMMENT_ERROR,
-    data: err
-  }
-}
+    data: err,
+  };
+};
 
 const putCommentAccepted = (comment) => {
   return $.ajax({
     method: "PUT",
     url: "/api/v3/comments",
-    data: Object.assign(comment, {mod: 1})
-  })
-}
+    data: Object.assign(comment, { mod: 1 }),
+  });
+};
 
 export const changeCommentStatusToAccepted = (comment) => {
   comment.active = true;
   return (dispatch) => {
-    dispatch(optimisticCommentAccepted(comment))
+    dispatch(optimisticCommentAccepted(comment));
     return putCommentAccepted(comment).then(
-      res => {
+      (res) => {
         dispatch(acceptCommentSuccess(res));
         dispatch(populateAllCommentStores(comment.conversation_id));
       },
-      err => dispatch(acceptCommentError(err))
-    )
-  }
-}
+      (err) => dispatch(acceptCommentError(err))
+    );
+  };
+};
 
 /* moderator clicked reject comment */
 
 const optimisticCommentRejected = (comment) => {
   return {
     type: REJECT_COMMENT,
-    comment: comment
-  }
-}
+    comment: comment,
+  };
+};
 
 const rejectCommentSuccess = (data) => {
   return {
     type: REJECT_COMMENT_SUCCESS,
-    data: data
-  }
-}
+    data: data,
+  };
+};
 
 const rejectCommentError = (err) => {
   return {
     type: REJECT_COMMENT_ERROR,
-    data: err
-  }
-}
+    data: err,
+  };
+};
 
 const putCommentRejected = (comment) => {
-
   return $.ajax({
     method: "PUT",
     url: "/api/v3/comments",
-    data: Object.assign(comment, {mod: -1})
-  })
-}
+    data: Object.assign(comment, { mod: -1 }),
+  });
+};
 
 export const changeCommentStatusToRejected = (comment) => {
   return (dispatch) => {
-    dispatch(optimisticCommentRejected())
+    dispatch(optimisticCommentRejected());
     return putCommentRejected(comment).then(
       (res) => {
         dispatch(rejectCommentSuccess(res));
         dispatch(populateAllCommentStores(comment.conversation_id));
       },
-      err => dispatch(rejectCommentError(err))
-    )
+      (err) => dispatch(rejectCommentError(err))
+    );
   };
 };
 
@@ -1272,42 +1273,42 @@ export const changeCommentStatusToRejected = (comment) => {
 const optimisticCommentIsMetaChanged = (comment) => {
   return {
     type: COMMENT_IS_META,
-    comment: comment
-  }
-}
+    comment: comment,
+  };
+};
 
 const commentIsMetaChangeSuccess = (data) => {
   return {
     type: COMMENT_IS_META_SUCCESS,
-    data: data
-  }
-}
+    data: data,
+  };
+};
 
 const commentIsMetaChangeError = (err) => {
   return {
     type: COMMENT_IS_META_ERROR,
-    data: err
-  }
-}
+    data: err,
+  };
+};
 
 const putCommentCommentIsMetaChange = (comment, is_meta) => {
   return $.ajax({
     method: "PUT",
     url: "/api/v3/comments",
-    data: Object.assign(comment, {is_meta: is_meta})
-  })
-}
+    data: Object.assign(comment, { is_meta: is_meta }),
+  });
+};
 
 export const changeCommentCommentIsMeta = (comment, is_meta) => {
   return (dispatch) => {
-    dispatch(optimisticCommentIsMetaChanged())
+    dispatch(optimisticCommentIsMetaChanged());
     return putCommentCommentIsMetaChange(comment, is_meta).then(
       (res) => {
         dispatch(commentIsMetaChangeSuccess(res));
         dispatch(populateAllCommentStores(comment.conversation_id));
       },
-      err => dispatch(commentIsMetaChangeError(err))
-    )
+      (err) => dispatch(commentIsMetaChangeError(err))
+    );
   };
 };
 
@@ -1315,21 +1316,21 @@ export const changeCommentCommentIsMeta = (comment, is_meta) => {
 
 const requestParticipants = () => {
   return {
-    type: REQUEST_PARTICIPANTS
+    type: REQUEST_PARTICIPANTS,
   };
 };
 
 const receiveParticipants = (data) => {
   return {
     type: RECEIVE_PARTICIPANTS,
-    data: data
+    data: data,
   };
 };
 
 const participantsFetchError = (err) => {
   return {
     type: PARTICIPANTS_FETCH_ERROR,
-    data: err
+    data: err,
   };
 };
 
@@ -1347,7 +1348,6 @@ export const populateParticipantsStore = (conversation_id) => {
   };
 };
 
-
 /* request default participants for ptpt moderation view */
 
 const requestDefaultParticipants = () => {
@@ -1359,30 +1359,30 @@ const requestDefaultParticipants = () => {
 const receiveDefaultParticipants = (data) => {
   return {
     type: RECEIVE_DEFAULT_PARTICIPANTS,
-    data: data
+    data: data,
   };
 };
 
 const defaultParticipantFetchError = (err) => {
   return {
     type: DEFAULT_PARTICIPANTS_FETCH_ERROR,
-    data: err
-  }
-}
+    data: err,
+  };
+};
 
 const fetchDefaultParticipants = (conversation_id) => {
   return $.get("/api/v3/ptptois?mod=0&conversation_id=" + conversation_id);
-}
+};
 
 export const populateDefaultParticipantStore = (conversation_id) => {
   return (dispatch) => {
-    dispatch(requestDefaultParticipants())
+    dispatch(requestDefaultParticipants());
     return fetchDefaultParticipants(conversation_id).then(
-      res => dispatch(receiveDefaultParticipants(res)),
-      err => dispatch(defaultParticipantFetchError(err))
-    )
-  }
-}
+      (res) => dispatch(receiveDefaultParticipants(res)),
+      (err) => dispatch(defaultParticipantFetchError(err))
+    );
+  };
+};
 
 /* request featured participants for ptpt moderation view */
 
@@ -1395,30 +1395,30 @@ const requestFeaturedParticipants = () => {
 const receiveFeaturedParticipants = (data) => {
   return {
     type: RECEIVE_FEATURED_PARTICIPANTS,
-    data: data
+    data: data,
   };
 };
 
 const featuredParticipantFetchError = (err) => {
   return {
     type: FEATURED_PARTICIPANTS_FETCH_ERROR,
-    data: err
-  }
-}
+    data: err,
+  };
+};
 
 const fetchFeaturedParticipants = (conversation_id) => {
   return $.get("/api/v3/ptptois?mod=1&conversation_id=" + conversation_id);
-}
+};
 
 export const populateFeaturedParticipantStore = (conversation_id) => {
   return (dispatch) => {
-    dispatch(requestFeaturedParticipants())
+    dispatch(requestFeaturedParticipants());
     return fetchFeaturedParticipants(conversation_id).then(
-      res => dispatch(receiveFeaturedParticipants(res)),
-      err => dispatch(featuredParticipantFetchError(err))
-    )
-  }
-}
+      (res) => dispatch(receiveFeaturedParticipants(res)),
+      (err) => dispatch(featuredParticipantFetchError(err))
+    );
+  };
+};
 
 /* request hidden participants for ptpt moderation view */
 
@@ -1431,30 +1431,30 @@ const requestHiddenParticipants = () => {
 const receiveHiddenParticipants = (data) => {
   return {
     type: RECEIVE_HIDDEN_PARTICIPANTS,
-    data: data
+    data: data,
   };
 };
 
 const hiddenParticipantFetchError = (err) => {
   return {
     type: HIDDEN_PARTICIPANTS_FETCH_ERROR,
-    data: err
-  }
-}
+    data: err,
+  };
+};
 
 const fetchHiddenParticipants = (conversation_id) => {
   return $.get("/api/v3/ptptois?mod=-1&conversation_id=" + conversation_id);
-}
+};
 
 export const populateHiddenParticipantStore = (conversation_id) => {
   return (dispatch) => {
-    dispatch(requestHiddenParticipants())
+    dispatch(requestHiddenParticipants());
     return fetchHiddenParticipants(conversation_id).then(
-      res => dispatch(receiveHiddenParticipants(res)),
-      err => dispatch(hiddenParticipantFetchError(err))
-    )
-  }
-}
+      (res) => dispatch(receiveHiddenParticipants(res)),
+      (err) => dispatch(hiddenParticipantFetchError(err))
+    );
+  };
+};
 
 /* populate ALL stores todo/accept/reject/seed */
 
@@ -1464,131 +1464,130 @@ export const populateAllParticipantStores = (conversation_id) => {
       dispatch(populateDefaultParticipantStore(conversation_id)),
       dispatch(populateFeaturedParticipantStore(conversation_id)),
       dispatch(populateHiddenParticipantStore(conversation_id))
-    )
-  }
-}
-
+    );
+  };
+};
 
 /* moderator clicked feature ptpt */
 
 const optimisticFeatureParticipant = (participant) => {
   return {
     type: FEATURE_PARTICIPANT,
-    participant: participant
-  }
-}
+    participant: participant,
+  };
+};
 
 const featureParticipantSuccess = (data) => {
   return {
     type: FEATURE_PARTICIPANT_SUCCESS,
-    data: data
-  }
-}
+    data: data,
+  };
+};
 
 const featureParticipantError = (err) => {
   return {
     type: FEATURE_PARTICIPANT_ERROR,
-    data: err
-  }
-}
+    data: err,
+  };
+};
 
 const putFeatureParticipant = (participant) => {
   return $.ajax({
     method: "PUT",
     url: "/api/v3/ptptois",
-    data: Object.assign(participant, {mod: 1})
-  })
-}
+    data: Object.assign(participant, { mod: 1 }),
+  });
+};
 
 export const changeParticipantStatusToFeatured = (participant) => {
   return (dispatch) => {
-    dispatch(optimisticFeatureParticipant(participant))
+    dispatch(optimisticFeatureParticipant(participant));
     return putFeatureParticipant(participant).then(
-      res => dispatch(featureParticipantSuccess(res)),
-      err => dispatch(featureParticipantError(err))
-    )
-  }
-}
+      (res) => dispatch(featureParticipantSuccess(res)),
+      (err) => dispatch(featureParticipantError(err))
+    );
+  };
+};
 /* moderator clicked hide ptpt */
 
 const optimisticHideParticipant = (participant) => {
   return {
     type: FEATURE_PARTICIPANT,
-    participant: participant
-  }
-}
+    participant: participant,
+  };
+};
 
 const hideParticipantSuccess = (data) => {
   return {
     type: FEATURE_PARTICIPANT_SUCCESS,
-    data: data
-  }
-}
+    data: data,
+  };
+};
 
 const hideParticipantError = (err) => {
   return {
     type: FEATURE_PARTICIPANT_ERROR,
-    data: err
-  }
-}
+    data: err,
+  };
+};
 
 const putHideParticipant = (participant) => {
   return $.ajax({
     method: "PUT",
     url: "/api/v3/ptptois",
-    data: Object.assign(participant, {mod: -1})
-  })
-}
+    data: Object.assign(participant, { mod: -1 }),
+  });
+};
 
 export const changeParticipantStatusToHidden = (participant) => {
   return (dispatch) => {
-    dispatch(optimisticHideParticipant(participant))
+    dispatch(optimisticHideParticipant(participant));
     return putHideParticipant(participant).then(
-      res => dispatch(hideParticipantSuccess(res)),
-      err => dispatch(hideParticipantError(err))
-    )
-  }
-}
+      (res) => dispatch(hideParticipantSuccess(res)),
+      (err) => dispatch(hideParticipantError(err))
+    );
+  };
+};
 
 /* moderator clicked unmoderate ptpt */
 const optimisticUnmoderateParticipant = (participant) => {
   return {
     type: FEATURE_PARTICIPANT,
-    participant: participant
-  }
-}
+    participant: participant,
+  };
+};
 
 const unmoderateParticipantSuccess = (data) => {
   return {
     type: FEATURE_PARTICIPANT_SUCCESS,
-    data: data
-  }
-}
+    data: data,
+  };
+};
 
 const unmoderateParticipantError = (err) => {
   return {
     type: FEATURE_PARTICIPANT_ERROR,
-    data: err
-  }
-}
+    data: err,
+  };
+};
 
 const putUnmoderateParticipant = (participant) => {
   return $.ajax({
     method: "PUT",
     url: "/api/v3/ptptois",
-    data: Object.assign(participant, {mod: 0})
-  })
-}
+    data: Object.assign(participant, { mod: 0 }),
+  });
+};
 
 export const changeParticipantStatusToUnmoderated = (participant) => {
   return (dispatch) => {
-    dispatch(optimisticUnmoderateParticipant(participant))
+    dispatch(optimisticUnmoderateParticipant(participant));
     return putUnmoderateParticipant(participant).then(
-      res => dispatch(hideParticipantSuccess(res)),
-      err => dispatch(hideParticipantError(err))
-    )
-  }
-}
+      (res) => dispatch(hideParticipantSuccess(res)),
+      (err) => dispatch(hideParticipantError(err))
+    );
+  };
+};
 
 /* request conversation stats */
 
@@ -1601,49 +1600,52 @@ const requestConversationStats = () => {
 const receiveConversationStats = (data) => {
   return {
     type: RECEIVE_CONVERSATION_STATS,
-    data: data
+    data: data,
   };
 };
 
 const conversationStatsFetchError = (err) => {
   return {
     type: CONVERSATION_STATS_FETCH_ERROR,
-    data: err
-  }
-}
+    data: err,
+  };
+};
 
 const fetchConversationStats = (conversation_id, until) => {
-  console.log("calling", conversation_id, until);
-  return $.get("/api/v3/conversationStats?conversation_id=" + conversation_id + (until ? "&until="+until : ""));
-}
+  return $.get(
+    "/api/v3/conversationStats?conversation_id=" +
+      conversation_id +
+      (until ? "&until=" + until : "")
+  );
+};
 
 export const populateConversationStatsStore = (conversation_id, until) => {
   return (dispatch) => {
-    dispatch(requestConversationStats())
+    dispatch(requestConversationStats());
     return fetchConversationStats(conversation_id, until).then(
-      res => dispatch(receiveConversationStats(res)),
-      err => dispatch(conversationStatsFetchError(err))
-    )
-  }
-}
+      (res) => dispatch(receiveConversationStats(res)),
+      (err) => dispatch(conversationStatsFetchError(err))
+    );
+  };
+};
 
 /* data export */
 
 const dataExportStarted = () => {
   return {
-    type: DATA_EXPORT_STARTED
+    type: DATA_EXPORT_STARTED,
   };
 };
 
 const dataExportSuccess = () => {
   return {
-    type: DATA_EXPORT_SUCCESS
+    type: DATA_EXPORT_SUCCESS,
   };
 };
 
 const dataExportError = () => {
   return {
-    type: DATA_EXPORT_ERROR
+    type: DATA_EXPORT_ERROR,
   };
 };
 
@@ -1656,14 +1658,14 @@ const dataExportGet = (conversation_id, format, unixTimestamp, untilEnabled) => 
     url += `&unixTimestamp=${unixTimestamp}`;
   }
   return $.get(url);
-}
+};
 
 export const startDataExport = (conversation_id, format, unixTimestamp, untilEnabled) => {
   return (dispatch) => {
-    dispatch(dataExportStarted() )
+    dispatch(dataExportStarted());
     return dataExportGet(conversation_id, format, unixTimestamp, untilEnabled).then(
-      res => dispatch(dataExportSuccess(res)),
-      err => dispatch(dataExportError(err))
-    )
-  }
-}
+      (res) => dispatch(dataExportSuccess(res)),
+      (err) => dispatch(dataExportError(err))
+    );
+  };
+};
