@@ -1,33 +1,53 @@
-// Copyright (C) 2012-present, The Authors. This program is free software: you can redistribute it and/or  modify it under the terms of the GNU Affero General Public License, version 3, as published by the Free Software Foundation. This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for more details. You should have received a copy of the GNU Affero General Public License along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
 import React from "react";
-import { browserHistory } from "react-router";
-import { connect } from "react-redux";
+import Layout from "./lander-layout";
+import { Heading, Box, Text, Link } from "theme-ui";
+import ExploreKnowledgeBase from "./exploreKnowledgeBase";
+import Press from "./press";
 
-import PolisNet from "../../util/net";
-import strings from "../../strings/strings";
-import { doSignin, doFacebookSignin } from "../../actions";
+const Index = () => {
+  return (
+    <Layout>
+      <Heading as="h1" sx={{ my: [4, null, 5], fontSize: [6, null, 7] }}>
+        Input Crowd, Output Meaning
+      </Heading>
+      <Heading
+        as="h3"
+        sx={{
+          fontSize: [3, null, 4],
+          lineHeight: "body",
+          mb: [4, null, 5],
+        }}
+      >
+        Polis is a real-time system for gathering, analyzing and understanding what large groups of
+        people think in their own words, enabled by advanced statistics and machine learning.
+      </Heading>
+      <Box sx={{ mb: [4, null, 5] }}>
+        <Text>
+          Polis has been used all over the world by governments, academics, independent media and
+          citizens, and is completely open source.
+        </Text>
+      </Box>
+      <Heading as="h3" sx={{ fontSize: [4], lineHeight: "body", mb: [2, null, 3] }}>
+        Get Started
+      </Heading>
+      <Box sx={{ mb: [4, null, 5] }}>
+        <Link href="/createuser">Sign up</Link>
+        {" or "}
+        <Link href="/signin">Sign in</Link>
+      </Box>
+      <Press />
+      <ExploreKnowledgeBase />
+      <Heading as="h3" sx={{ fontSize: [4], lineHeight: "body", my: [2, null, 3] }}>
+        Contribute
+      </Heading>
+      <Box sx={{ mb: [4, null, 5] }}>
+        Explore the code and join the community{" "}
+        <Link target="_blank" href="https://github.com/pol-is/">
+          on Github
+        </Link>
+      </Box>
+    </Layout>
+  );
+};
 
-import CTA from "./CTA/CTA";
-import Nav from "../App/Nav/Nav";
-import FeatureSection from "./Features/FeatureSection";
-import Footer from "../App/Footer/Footer";
-import FooterData from '../../strings/footer'
-import Hero from "./Hero/Hero";
-import Trust from "./Trust/Trust";
-
-@connect()
-class Home extends React.Component {
-  render() {
-    return <main>
-        <Nav />
-        <Hero headline="Know what your organization is thinking" subheadline="Polis helps organizations understand themselves by visualizing what their members think. Get a clear picture of all the points of view to help move a conversation forward." className="page-header mt6 mt5-ns" />
-        <Trust headline="You’re in good company" subheadline="Polis is trusted by governments, universities, non-profits, movements, and large organizations." />
-        <FeatureSection />
-        <CTA />
-        <Footer social={FooterData.footer.social} content={FooterData.footer.groups} data={FooterData.footer} />
-      </main>;
-  }
-}
-
-export default Home;
+export default Index;
