@@ -5,10 +5,12 @@ var s = {};
 // login/createuser strings
 s.polis_err_param_parse_failed_email = "Please enter a valid email address.";
 s.polis_err_reg_bad_email = "Please enter a valid email address.";
-s.polis_err_param_parse_failed_password = "Please enter a valid password."
-s.polis_err_login_unknown_user_or_password_noresults = "Login failed: invalid username/password combination.";
+s.polis_err_param_parse_failed_password = "Please enter a valid password.";
+s.polis_err_login_unknown_user_or_password_noresults =
+  "Login failed: invalid username/password combination.";
 s.polis_err_login_unknown_user_or_password = "Login failed: invalid username/password combination.";
-s.polis_err_reg_user_with_that_email_exists = "Email address already in use, Try logging in instead.";
+s.polis_err_reg_user_with_that_email_exists =
+  "Email address already in use, Try logging in instead.";
 s.polis_err_reg_need_name = "Please include your name.";
 
 s.polis_err_post_comment_duplicate = "Error posting: This comment already exists!";
@@ -16,10 +18,11 @@ s.waitinglist_add_success = "You've been added to the waiting list! We'll be in 
 
 s.polis_err_fetching_tweet = "Error fetching tweet. Expected a URL to a twitter tweet.";
 
-s.share_but_no_comments_warning = "This conversation has no comments. We recommend you add a few comments before inviting participants. This will help participants get started. In the menu to the left, go to 'comments' and then 'seed'.";
-s.share_but_no_visible_comments_warning = "This conversation has no visible comments. We recommend you add a few comments (or moderate the comments that exist) before inviting participants, since this will help them understand what kind of comments they should submit.";
+s.share_but_no_comments_warning =
+  "This conversation has no comments. We recommend you add a few comments before inviting participants. This will help participants get started. Go to 'Configure' and then 'Seed Comments'.";
+s.share_but_no_visible_comments_warning =
+  "This conversation has no visible comments. We recommend you add a few comments (or moderate the comments that exist) before inviting participants, since this will help them understand what kind of comments they should submit.";
 
 s.no_permission = "Your account does not have the permissions to view this page.";
-
 
 export default s;
